+++ common/src/main/java/net/untitledduckmod/client/model/GooseModel.java
@@ -1,11 +1,11 @@
 package net.untitledduckmod.client.model;
 
+import com.geckolib.renderer.base.GeoRenderState;
 import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 import net.minecraft.resources.Identifier;
 import net.untitledduckmod.DuckMod;
 import net.untitledduckmod.common.entity.GooseEntity;
 import net.untitledduckmod.common.entity.WaterfowlEntity;
-import software.bernie.geckolib.renderer.base.GeoRenderState;
 
 import java.util.Objects;