+++ common/src/main/java/net/untitledduckmod/common/platform/Services.java
@@ -12,7 +12,7 @@ public class Services {
     public static final IUntitledConfig CONFIG = load(IUntitledConfig.class);
 
     public static <T> T load(Class<T> clazz) {
-        final T loadedService = ServiceLoader.load(clazz)
+        final T loadedService = ServiceLoader.load(clazz, Services.class.getClassLoader())
                 .findFirst()
                 .orElseThrow(() -> new NullPointerException("Failed to load service for " + clazz.getName()));
         return loadedService;