+++ fabric/src/main/java/net/untitledduckmod/common/init/fabric/FabricModPotions.java
@@ -1,12 +1,12 @@
 package net.untitledduckmod.common.init.fabric;
 
-import net.fabricmc.fabric.api.registry.FabricBrewingRecipeRegistryBuilder;
+import net.fabricmc.fabric.api.registry.FabricPotionBrewingBuilder;
 import net.untitledduckmod.common.init.ModPotions;
 
 public class FabricModPotions extends ModPotions {
 
     public FabricModPotions() {
         super();
-        FabricBrewingRecipeRegistryBuilder.BUILD.register(ModPotions::registerRecipes);
+        FabricPotionBrewingBuilder.BUILD.register(ModPotions::registerRecipes);
     }
 }