+++ docs/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version ["v1.5.5"] - 2026-4-04
+
+### Features
+- The goose can now use the spear
+
 ## Version ["v1.5.3"] - 2026-3-14
 
 ### Bug fixes